style(var_expand): ファイル分割を実行
tokenリストを分割する処理をexpand_tokensへ
文字列を変数展開を変数展開する処理をexpanderへ
一つの変数を展開し、前後の文字列とつなぎ合わせる処理をexpand_varへ
