+++ includes/expander.h
@@ -6,7 +6,7 @@
 /*   By: totaisei <totaisei@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/21 08:24:01 by totaisei          #+#    #+#             */
-/*   Updated: 2021/02/22 11:52:50 by totaisei         ###   ########.fr       */
+/*   Updated: 2021/02/22 12:11:07 by totaisei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@ typedef struct	s_expander
 }				t_expander;
 
 
-void expand_tokens(t_token **tokens);
-void expander_init(t_expander *exper, char *input);
+void	expand_tokens(t_token **tokens);
+char	*create_expanded_str(const char *str,  t_token_state state);
+char	*expand_env_var(char *input);
 #endif
